+++ CHANGES.md
@@ -1,7 +1,7 @@
 v4.1.0
 ------
 
-- Add support for Alpine 3.7
+- Add support for Alpine 3.7 and Ubuntu 18.04.
 - Alter the individual compiler containers to omit the patch version
   from the name. They will always have the latest patch version for CI.